pyproject.toml
[project]
name = "analytic-nl-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.23.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "llama-cpp-python>=0.3.16",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "sqlparse>=0.5.4",
]

[dependency-groups]
dev = [
    "deptry>=0.24.0",
]
